just compute the line count data and dump to json
